refactor: update buildCanonicalUrl to recieve canonical url from Sanity
