+++ packages/osc-ecommerce/app/routes/blog/$slug.tsx
@@ -10,7 +10,7 @@ import { POST_QUERY } from '~/queries/sanity/post';
 import type { SanityPage } from '~/types/sanity';
 import Module from '~/components/Module';
 import type { module } from '~/types/sanity';
-import { buildCanonicalUrl } from '~/utils/buildCanonicalUrl';
+import { buildCanonicalUrl } from '~/utils/metaTags/buildCanonicalUrl';
 
 interface PageData {
     page: SanityPage;
@@ -32,7 +32,10 @@ export const loader: LoaderFunction = async ({ request, params }) => {
     }
 
     const { page: post, isPreview }: PageData = data;
-    const canonicalUrl = buildCanonicalUrl(request);
+    const canonicalUrl = buildCanonicalUrl({
+        canonical: post?.seo?.canonicalUrl,
+        request
+    });
 
     return json({
         post,
