+++ packages/osc-ecommerce/app/routes/collections/$slug.tsx
@@ -10,7 +10,7 @@ import { COLLECTION_QUERY } from '~/queries/sanity/collection';
 import type { SanityPage } from '~/types/sanity';
 import Module from '~/components/Module';
 import type { module } from '~/types/sanity';
-import { buildCanonicalUrl } from '~/utils/buildCanonicalUrl';
+import { buildCanonicalUrl } from '~/utils/metaTags/buildCanonicalUrl';
 
 interface PageData {
     page: SanityPage;
@@ -32,7 +32,10 @@ export const loader: LoaderFunction = async ({ request, params }) => {
     }
 
     const { page: collection, isPreview }: PageData = data;
-    const canonicalUrl = buildCanonicalUrl(request);
+    const canonicalUrl = buildCanonicalUrl({
+        canonical: collection?.seo?.canonicalUrl,
+        request
+    });
 
     return json({
         collection,
